# PERF: ArrowDtype.__eq__