--- a/pandas/core/dtypes/dtypes.py
+++ b/pandas/core/dtypes/dtypes.py
@@ -2046,6 +2046,15 @@ class ArrowDtype(StorageExtensionDtype):
     def __repr__(self) -> str:
         return self.name
 
+    def __hash__(self) -> int:
+        # make myself hashable
+        return hash(str(self))
+
+    def __eq__(self, other: Any) -> bool:
+        if not isinstance(other, type(self)):
+            return super().__eq__(other)
+        return self.pyarrow_dtype == other.pyarrow_dtype
+
     @property
     def type(self):
         """
@@ -2104,6 +2113,12 @@ class ArrowDtype(StorageExtensionDtype):
         """
         A string identifying the data type.
         """
+        # try:
+        #     return self._cache_dtype_names[self.pyarrow_dtype]
+        # except KeyError:
+        #     name = f"{str(self.pyarrow_dtype)}[{self.storage}]"
+        #     self._cache_dtype_names[self.pyarrow_dtype] = name
+        #     return name
         return f"{str(self.pyarrow_dtype)}[{self.storage}]"
 
     @cache_readonly