Micro-perf improvement:

```
import pandas as pd
import pyarrow as pa

a = pd.ArrowDtype(pa.float64())
b = pd.ArrowDtype(pa.float64())

%timeit a == b

# 2.86 µs ± 214 ns per loop (mean ± std. dev. of 7 runs, 100,000 loops each)   -> main
# 473 ns ± 15.2 ns per loop (mean ± std. dev. of 7 runs, 1,000,000 loops each) -> PR
```